refactor(identity): domain enum 패키지 분리 #48
