+++ systems/identity/identity-domain/src/main/kotlin/hs/kr/entrydsm/identity/domain/enum/SignupType.kt
@@ -1,4 +1,4 @@
-package hs.kr.entrydsm.identity.domain
+package hs.kr.entrydsm.identity.domain.enum
 
 enum class SignupType {
     SELF,
